requirements.txt
# ── Deep learning ─────────────────────────────────────────────────────────────
torch==2.2.2
transformers==4.40.2
datasets==2.19.2
tokenizers==0.19.1
accelerate==0.30.1

# ── Evaluation / metrics ──────────────────────────────────────────────────────
scikit-learn==1.4.2
matplotlib==3.9.0
numpy==1.26.4

# ── Experiment tracking ───────────────────────────────────────────────────────
mlflow==2.13.2

# ── LlamaIndex (document pipeline + RAG) ─────────────────────────────────────
llama-index-core
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface

# ── Embeddings ────────────────────────────────────────────────────────────────
sentence-transformers==3.0.1

# ── Vector database ───────────────────────────────────────────────────────────
chromadb==0.5.5

# ── LangGraph / LangChain (agent orchestration) ───────────────────────────────
langchain-core
langchain-community
langchain-ollama
langgraph

# ── Caching ───────────────────────────────────────────────────────────────────
redis==5.0.8
python-dotenv==1.0.1

# ── Market data ───────────────────────────────────────────────────────────────
yfinance==0.2.43

# ── API server ────────────────────────────────────────────────────────────────
fastapi==0.111.1
uvicorn[standard]==0.30.1
pydantic==2.7.4
httpx==0.27.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.2

# ── Linting ───────────────────────────────────────────────────────────────────
flake8==7.1.0
